# backend/public/swagger-spec.json
{"openapi":"3.0.0","info":{"title":"Online Coding Platform API","version":"1.0.0","description":"API documentation for Online Coding Platform","contact":{"name":"Huy-VNNIC","email":"support@codingplatform.com"},"license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"}},"servers":[{"url":"/api","description":"Development server"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"User":{"type":"object","required":["name","email","password"],"properties":{"id":{"type":"integer","description":"User ID"},"name":{"type":"string","description":"User name"},"email":{"type":"string","format":"email","description":"User email"},"role":{"type":"string","enum":["student","instructor","support","marketing","developer"],"description":"User role"},"verified":{"type":"boolean","description":"Email verification status"},"created_at":{"type":"string","format":"date-time","description":"User creation timestamp"}}},"Student":{"type":"object","properties":{"user_id":{"type":"integer","description":"User ID"},"dashboard_data":{"type":"object","description":"Dashboard data as JSON"}}},"Resource":{"type":"object","properties":{"id":{"type":"integer","description":"Resource ID"},"title":{"type":"string","description":"Resource title"},"type":{"type":"string","enum":["video","slide","ebook","code"],"description":"Resource type"},"language":{"type":"string","description":"Programming language"},"url":{"type":"string","description":"Resource URL"},"created_by":{"type":"integer","description":"Creator user ID"},"created_at":{"type":"string","format":"date-time","description":"Resource creation timestamp"}}},"Submission":{"type":"object","properties":{"id":{"type":"integer","description":"Submission ID"},"student_id":{"type":"integer","description":"Student user ID"},"resource_id":{"type":"integer","description":"Resource ID"},"code":{"type":"string","description":"Submitted code"},"result":{"type":"string","description":"Execution result"},"status":{"type":"string","enum":["pending","success","failed"],"description":"Submission status"},"submitted_at":{"type":"string","format":"date-time","description":"Submission timestamp"}}},"Response":{"type":"object","properties":{"success":{"type":"boolean","description":"Success status"},"message":{"type":"string","description":"Response message"},"data":{"type":"object","description":"Response data"},"timestamp":{"type":"string","format":"date-time","description":"Response timestamp"}}},"Error":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string"},"data":{"type":"object","nullable":true},"timestamp":{"type":"string","format":"date-time"}}}}},"paths":{"/api/auth/register":{"post":{"summary":"Register a new student","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["name","email","password"],"properties":{"name":{"type":"string","example":"John Doe"},"email":{"type":"string","format":"email","example":"john@example.com"},"password":{"type":"string","format":"password","example":"securepassword123"}}}}}},"responses":{"201":{"description":"Registration successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Email already in use","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/auth/verify-email":{"post":{"summary":"Verify user email","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["token"],"properties":{"token":{"type":"string","example":"5f4dcc3b5aa765d61d8327deb882cf99"}}}}}},"responses":{"200":{"description":"Email verified successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"}}}},"400":{"description":"Invalid or expired token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/auth/login":{"post":{"summary":"Login user","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email","example":"john@example.com"},"password":{"type":"string","format":"password","example":"securepassword123"}}}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Login successful"},"data":{"type":"object","properties":{"token":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"user":{"type":"object","properties":{"id":{"type":"integer","example":1},"name":{"type":"string","example":"John Doe"},"email":{"type":"string","example":"john@example.com"},"role":{"type":"string","example":"student"}}}}},"timestamp":{"type":"string","format":"date-time"}}}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"403":{"description":"Account not verified","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/auth/forgot-password":{"post":{"summary":"Request password reset","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email"],"properties":{"email":{"type":"string","format":"email","example":"john@example.com"}}}}}},"responses":{"200":{"description":"Password reset email sent","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/auth/reset-password":{"post":{"summary":"Reset password","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["token","password"],"properties":{"token":{"type":"string","example":"5f4dcc3b5aa765d61d8327deb882cf99"},"password":{"type":"string","format":"password","example":"newpassword123"}}}}}},"responses":{"200":{"description":"Password reset successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"}}}},"400":{"description":"Invalid or expired token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/resources":{"get":{"summary":"Get all learning resources","tags":["Resources"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"type","schema":{"type":"string","enum":["video","slide","ebook","code"]},"description":"Filter resources by type"}],"responses":{"200":{"description":"Resources retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Resources retrieved successfully"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Resource"}},"timestamp":{"type":"string","format":"date-time"}}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/resources/search":{"get":{"summary":"Search resources","tags":["Resources"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"keyword","required":true,"schema":{"type":"string"},"description":"Search keyword"}],"responses":{"200":{"description":"Search results","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Search results"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Resource"}},"timestamp":{"type":"string","format":"date-time"}}}}}},"400":{"description":"Search keyword is required","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/resources/{id}":{"get":{"summary":"Get resource by ID","tags":["Resources"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"},"description":"Resource ID"}],"responses":{"200":{"description":"Resource retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Resource retrieved successfully"},"data":{"$ref":"#/components/schemas/Resource"},"timestamp":{"type":"string","format":"date-time"}}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Resource not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/student/profile":{"get":{"summary":"Get student profile","tags":["Student"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Profile retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Profile retrieved successfully"},"data":{"allOf":[{"$ref":"#/components/schemas/User"},{"type":"object","properties":{"dashboardData":{"type":"object"}}}]},"timestamp":{"type":"string","format":"date-time"}}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Profile not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/student/dashboard":{"get":{"summary":"Get student dashboard","tags":["Student"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Dashboard retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Dashboard retrieved successfully"},"data":{"type":"object","properties":{"progress":{"type":"number","example":75},"completedResources":{"type":"array","items":{"type":"integer"}},"achievements":{"type":"array","items":{"type":"object"}}}},"timestamp":{"type":"string","format":"date-time"}}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"put":{"summary":"Update student dashboard","tags":["Student"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["dashboardData"],"properties":{"dashboardData":{"type":"object","example":{"preferences":{"theme":"dark","notifications":true},"favoriteResources":[1,3,5]}}}}}}},"responses":{"200":{"description":"Dashboard updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/submissions":{"post":{"summary":"Submit code for execution","tags":["Submissions"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["resourceId","code","language"],"properties":{"resourceId":{"type":"integer","example":1},"code":{"type":"string","example":"function square(x) { return x * x; }"},"language":{"type":"string","example":"javascript"}}}}}},"responses":{"201":{"description":"Code submitted successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Code submitted successfully"},"data":{"type":"object","properties":{"submissionId":{"type":"integer","example":42},"status":{"type":"string","example":"pending"},"message":{"type":"string","example":"Your code is being processed. Check back soon for results."}}},"timestamp":{"type":"string","format":"date-time"}}}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Resource not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"get":{"summary":"Get all submissions for a student","tags":["Submissions"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Submissions retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Submissions retrieved successfully"},"data":{"type":"array","items":{"allOf":[{"$ref":"#/components/schemas/Submission"},{"type":"object","properties":{"resource_title":{"type":"string","example":"JavaScript Basics"}}}]}},"timestamp":{"type":"string","format":"date-time"}}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/submissions/{id}":{"get":{"summary":"Get submission status","tags":["Submissions"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"},"description":"Submission ID"}],"responses":{"200":{"description":"Submission retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Submission retrieved successfully"},"data":{"$ref":"#/components/schemas/Submission"},"timestamp":{"type":"string","format":"date-time"}}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Submission not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/submissions/resource/{resourceId}":{"get":{"summary":"Get submissions for a specific resource","tags":["Submissions"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"resourceId","required":true,"schema":{"type":"integer"},"description":"Resource ID"}],"responses":{"200":{"description":"Resource submissions retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Resource submissions retrieved successfully"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Submission"}},"timestamp":{"type":"string","format":"date-time"}}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}}},"tags":[]}